+++ opener/src/lib.rs
@@ -25,8 +25,19 @@
     unused_qualifications
 )]
 
+#[cfg(not(any(target_os = "windows", target_os = "macos")))]
+mod linux_and_more;
+#[cfg(target_os = "macos")]
+mod macos;
 #[cfg(target_os = "windows")]
-use crate::windows::open_sys;
+mod windows;
+
+#[cfg(not(any(target_os = "windows", target_os = "macos")))]
+use crate::linux_and_more::open as open_sys;
+#[cfg(target_os = "macos")]
+use crate::macos::open as open_sys;
+#[cfg(target_os = "windows")]
+use crate::windows::open as open_sys;
 
 use std::{
     error::Error,
@@ -35,6 +46,22 @@ use std::{
     io,
     process::ExitStatus,
 };
+
+/// Opens a file or link with the system default program.
+///
+/// Note that a path like "rustup.rs" could potentially refer to either a file or a website. If you
+/// want to open the website, you should add the "http://" prefix, for example.
+///
+/// Also note that a result of `Ok(())` just means a way of opening the path was found, and no error
+/// occurred as a direct result of opening the path. Errors beyond that point aren't caught. For
+/// example, `Ok(())` would be returned even if a file was opened with a program that can't read the
+/// file, or a dead link was opened in a browser.
+pub fn open<P>(path: P) -> Result<(), OpenError>
+where
+    P: AsRef<OsStr>,
+{
+    open_sys(path.as_ref())
+}
 
 /// An error type representing the failure to open a path. Possibly returned by the [`open`]
 /// function.
@@ -92,132 +119,5 @@ impl Error for OpenError {
             OpenError::Io(inner) => Some(inner),
             OpenError::ExitStatus { .. } => None,
         }
-    }
-}
-
-impl From<io::Error> for OpenError {
-    fn from(err: io::Error) -> Self {
-        OpenError::Io(err)
-    }
-}
-
-/// Opens a file or link with the system default program.
-///
-/// Note that a path like "rustup.rs" could potentially refer to either a file or a website. If you
-/// want to open the website, you should add the "http://" prefix, for example.
-///
-/// Also note that a result of `Ok(())` just means a way of opening the path was found, and no error
-/// occurred as a direct result of opening the path. Errors beyond that point aren't caught. For
-/// example, `Ok(())` would be returned even if a file was opened with a program that can't read the
-/// file, or a dead link was opened in a browser.
-pub fn open<P>(path: P) -> Result<(), OpenError>
-where
-    P: AsRef<OsStr>,
-{
-    open_sys(path.as_ref())
-}
-
-#[cfg(target_os = "windows")]
-mod windows {
-    use super::OpenError;
-    use std::{ffi::OsStr, io, os::windows::ffi::OsStrExt, ptr};
-    use winapi::{ctypes::c_int, um::shellapi::ShellExecuteW};
-
-    pub fn open_sys(path: &OsStr) -> Result<(), OpenError> {
-        const SW_SHOW: c_int = 5;
-
-        let path = convert_path(path)?;
-        let operation: Vec<u16> = OsStr::new("open\0").encode_wide().collect();
-        let result = unsafe {
-            ShellExecuteW(
-                ptr::null_mut(),
-                operation.as_ptr(),
-                path.as_ptr(),
-                ptr::null(),
-                ptr::null(),
-                SW_SHOW,
-            )
-        };
-        if result as c_int > 32 {
-            Ok(())
-        } else {
-            Err(io::Error::last_os_error().into())
-        }
-    }
-
-    fn convert_path(path: &OsStr) -> io::Result<Vec<u16>> {
-        let mut maybe_result: Vec<u16> = path.encode_wide().collect();
-        if maybe_result.iter().any(|&u| u == 0) {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                "path contains NUL byte(s)",
-            ));
-        }
-        maybe_result.push(0);
-        Ok(maybe_result)
-    }
-}
-
-#[cfg(target_os = "macos")]
-fn open_sys(path: &OsStr) -> Result<(), OpenError> {
-    open_not_windows("open", path, &[], None, "open")
-}
-
-#[cfg(not(any(target_os = "windows", target_os = "macos")))]
-fn open_sys(path: &OsStr) -> Result<(), OpenError> {
-    const XDG_OPEN_SCRIPT: &[u8] = include_bytes!("xdg-open");
-
-    open_not_windows(
-        "sh",
-        path,
-        &["-s"],
-        Some(XDG_OPEN_SCRIPT),
-        "xdg-open (internal)",
-    )
-}
-
-#[cfg(not(target_os = "windows"))]
-fn open_not_windows(
-    cmd: &str,
-    path: &OsStr,
-    extra_args: &[&str],
-    piped_input: Option<&[u8]>,
-    cmd_friendly_name: &'static str,
-) -> Result<(), OpenError> {
-    use std::{
-        io::{Read, Write},
-        process::{Command, Stdio},
-    };
-
-    let stdin_type = if piped_input.is_some() {
-        Stdio::piped()
-    } else {
-        Stdio::null()
-    };
-
-    let mut cmd = Command::new(cmd)
-        .args(extra_args)
-        .arg(path)
-        .stdin(stdin_type)
-        .stdout(Stdio::null())
-        .stderr(Stdio::piped())
-        .spawn()?;
-
-    if let Some(stdin) = cmd.stdin.as_mut() {
-        stdin.write_all(piped_input.unwrap())?;
-    }
-
-    let exit_status = cmd.wait()?;
-    if exit_status.success() {
-        Ok(())
-    } else {
-        let mut stderr = String::new();
-        cmd.stderr.as_mut().unwrap().read_to_string(&mut stderr)?;
-
-        Err(OpenError::ExitStatus {
-            cmd: cmd_friendly_name,
-            status: exit_status,
-            stderr,
-        })
     }
 }